refactor: utils tests
